+++ packages/vsce/CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to the "cics-extension-for-zowe" extension will be documente
 ## Recent Changes
 
 - BugFix: Fixed plex resource filtering to only filter regions. [#575](https://github.com/zowe/cics-for-zowe-client/issues/575)
+- Enhancement: Inquire program, Show Library and Inquire Transaction can open respective resources from Resource Inspector. [#449](https://github.com/zowe/cics-for-zowe-client/issues/449)
 
 ## `3.20.0`